각 역할별로 패키지 분리 및 클래스 JavaDoc 작성
