--- src/main/java/dooya/see/auth/application/AuthService.java
@@ -1,5 +1,0 @@
-package dooya.see.auth.application;
-
-public interface AuthService {
-    LoginResult login(LoginCommand command);
-}
