--- src/main/java/dooya/see/auth/application/AuthValidator.java
@@ -1,7 +1,0 @@
-package dooya.see.auth.application;
-
-import dooya.see.user.domain.User;
-
-public interface AuthValidator {
-    User validateEmailAndPassword(String email, String password);
-}
