--- src/main/java/dooya/see/auth/application/LoginCommand.java
@@ -1,10 +1,0 @@
-package dooya.see.auth.application;
-
-import lombok.Builder;
-
-@Builder
-public record LoginCommand(
-        String email,
-        String password
-) {
-}
