+++ src/main/java/dooya/see/auth/application/service/impl/AuthServiceImpl.java
@@ -1,5 +1,9 @@
-package dooya.see.auth.application;
+package dooya.see.auth.application.service.impl;
 
+import dooya.see.auth.application.dto.LoginCommand;
+import dooya.see.auth.application.dto.LoginResult;
+import dooya.see.auth.application.service.AuthService;
+import dooya.see.auth.application.service.AuthValidator;
 import dooya.see.auth.util.JwtUtil;
 import dooya.see.user.domain.User;
 import lombok.RequiredArgsConstructor;
