--- src/main/java/dooya/see/auth/presentation/LoginRequest.java
@@ -1,14 +1,0 @@
-package dooya.see.auth.presentation;
-
-import jakarta.validation.constraints.NotBlank;
-import lombok.Builder;
-
-@Builder(toBuilder = true)
-public record LoginRequest(
-        @NotBlank(message = "이메일은 비어 있을 수 없습니다")
-        String email,
-
-        @NotBlank(message = "비밀번호는 비어 있을 수 없습니다")
-        String password
-) {
-}
