--- src/main/java/dooya/see/auth/presentation/LoginResponse.java
@@ -1,13 +1,0 @@
-package dooya.see.auth.presentation;
-
-import lombok.Builder;
-
-@Builder
-public record LoginResponse(
-        String accessToken,
-        Long id,
-        String email,
-        String name,
-        String nickName
-) {
-}
