--- src/main/java/dooya/see/user/application/UserQueryService.java
@@ -1,5 +1,0 @@
-package dooya.see.user.application;
-
-public interface UserQueryService {
-    UserResult getUserByEmail(String email);
-}
