--- src/main/java/dooya/see/user/application/UserSignUpCommand.java
@@ -1,12 +1,0 @@
-package dooya.see.user.application;
-
-import lombok.Builder;
-
-@Builder
-public record UserSignUpCommand(
-        String email,
-        String name,
-        String password,
-        String nickName
-) {
-}
