--- src/main/java/dooya/see/user/application/UserSignUpService.java
@@ -1,5 +1,0 @@
-package dooya.see.user.application;
-
-public interface UserSignUpService {
-    UserResult userSignUp(UserSignUpCommand command);
-}
