--- src/main/java/dooya/see/user/application/UserUpdateCommand.java
@@ -1,6 +1,0 @@
-package dooya.see.user.application;
-
-public record UserUpdateCommand(
-        String nickName
-) {
-}
