--- src/main/java/dooya/see/user/application/UserUpdateService.java
@@ -1,5 +1,0 @@
-package dooya.see.user.application;
-
-public interface UserUpdateService {
-    UserResult updateNickName(String email, UserUpdateCommand command);
-}
