--- src/main/java/dooya/see/user/application/UserValidator.java
@@ -1,5 +1,0 @@
-package dooya.see.user.application;
-
-public interface UserValidator {
-    void validateDuplicateEmail(String email);
-}
