+++ src/main/java/dooya/see/user/application/dto/UserUpdateCommand.java
@@ -1,0 +1,17 @@
+package dooya.see.user.application.dto;
+
+/**
+ * {@code UserUpdateCommand}는 사용자 정보 업데이트를 위한
+ * 애플리케이션 계층의 커맨드 객체입니다.
+ *
+ * <p>이 객체는 사용자 닉네임 변경 요청 시 필요한 데이터를 캡슐화하여
+ * 서비스 계층에 전달하는 역할을 합니다.
+ *
+ * <p>불변 객체로 설계되어 안전하게 데이터를 전달할 수 있습니다.
+ *
+ * @author dooya
+ */
+public record UserUpdateCommand(
+        String nickName
+) {
+}
