+++ src/main/java/dooya/see/user/application/service/impl/UserSignUpServiceImpl.java
@@ -1,5 +1,10 @@
-package dooya.see.user.application;
+package dooya.see.user.application.service.impl;
 
+import dooya.see.user.application.dto.UserApplicationMapper;
+import dooya.see.user.application.dto.UserResult;
+import dooya.see.user.application.dto.UserSignUpCommand;
+import dooya.see.user.application.service.UserSignUpService;
+import dooya.see.user.application.service.UserValidator;
 import dooya.see.user.domain.User;
 import dooya.see.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
