+++ src/main/java/dooya/see/user/application/service/impl/UserUpdateServiceImpl.java
@@ -1,14 +1,17 @@
-package dooya.see.user.application;
+package dooya.see.user.application.service.impl;
 
 import dooya.see.common.exception.CustomException;
 import dooya.see.common.exception.ErrorCode;
+import dooya.see.user.application.dto.UserResult;
+import dooya.see.user.application.dto.UserUpdateCommand;
+import dooya.see.user.application.service.UserUpdateService;
 import dooya.see.user.domain.User;
 import dooya.see.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import static dooya.see.user.application.UserApplicationMapper.*;
+import static dooya.see.user.application.dto.UserApplicationMapper.*;
 
 @Service
 @RequiredArgsConstructor
