+++ src/main/java/dooya/see/user/application/service/impl/UserValidatorImpl.java
@@ -1,7 +1,8 @@
-package dooya.see.user.application;
+package dooya.see.user.application.service.impl;
 
 import dooya.see.common.exception.CustomException;
 import dooya.see.common.exception.ErrorCode;
+import dooya.see.user.application.service.UserValidator;
 import dooya.see.user.domain.User;
 import dooya.see.user.domain.UserRepository;
 import lombok.RequiredArgsConstructor;
