+++ src/main/java/dooya/see/user/presentation/UserPresentationMapper.java
@@ -1,9 +1,25 @@
 package dooya.see.user.presentation;
 
-import dooya.see.user.application.UserResult;
-import dooya.see.user.application.UserSignUpCommand;
-import dooya.see.user.application.UserUpdateCommand;
+import dooya.see.user.application.dto.UserResult;
+import dooya.see.user.application.dto.UserSignUpCommand;
+import dooya.see.user.application.dto.UserUpdateCommand;
 
+/**
+ * {@code UserPresentationMapper} 클래스는
+ * 프레젠테이션 계층에서 요청(Request) 및 응답(Response) 객체와
+ * 애플리케이션 계층의 커맨드(Command) 및 결과(Result) 객체 간 변환을 담당합니다.
+ *
+ * <p>주요 역할은 다음과 같습니다:
+ * <ul>
+ *     <li>클라이언트 요청 객체를 애플리케이션 커맨드 객체로 변환</li>
+ *     <li>애플리케이션 결과 객체를 클라이언트 응답 객체로 변환</li>
+ * </ul>
+ *
+ * <p>이를 통해 프레젠테이션 계층과 애플리케이션 계층 간의 의존성을 최소화하며,
+ * 객체 변환 책임을 명확히 분리합니다.
+ *
+ * @author dooya
+ */
 public class UserPresentationMapper {
 
     public static UserSignUpCommand toSignCommand(UserSignUpRequest request) {
