+++ src/test/java/dooya/see/auth/application/AuthServiceTest.java
@@ -1,5 +1,9 @@
 package dooya.see.auth.application;
 
+import dooya.see.auth.application.dto.LoginCommand;
+import dooya.see.auth.application.dto.LoginResult;
+import dooya.see.auth.application.service.impl.AuthServiceImpl;
+import dooya.see.auth.application.service.AuthValidator;
 import dooya.see.auth.util.JwtUtil;
 import dooya.see.common.AuthFixture;
 import dooya.see.common.UserFixture;
