+++ src/test/java/dooya/see/auth/presentation/AuthControllerTest.java
@@ -1,6 +1,7 @@
 package dooya.see.auth.presentation;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import dooya.see.auth.presentation.dto.LoginRequest;
 import dooya.see.common.UserFixture;
 import dooya.see.user.infrastructure.UserJpaRepository;
 import org.junit.jupiter.api.BeforeEach;
