+++ src/test/java/dooya/see/user/application/UserSignUpServiceTest.java
@@ -2,6 +2,10 @@ package dooya.see.user.application;
 
 import dooya.see.common.exception.CustomException;
 import dooya.see.common.exception.ErrorCode;
+import dooya.see.user.application.dto.UserResult;
+import dooya.see.user.application.dto.UserSignUpCommand;
+import dooya.see.user.application.service.impl.UserSignUpServiceImpl;
+import dooya.see.user.application.service.UserValidator;
 import dooya.see.user.domain.Role;
 import dooya.see.user.domain.User;
 import dooya.see.user.domain.UserRepository;
